moving back bipartite to bipartite.h
